requirements.txt
blinker==1.7.0
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
distlib==0.3.8
dnspython==2.5.0
filelock==3.14.0
Flask==3.0.2
Flask-SQLAlchemy==3.1.1
greenlet==3.0.3
idna==3.7
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.5
mysql-connector-python==8.4.0
passlib==1.7.4
platformdirs==4.2.1
pymongo==4.6.1
python-dotenv==1.0.1
requests==2.32.3
SQLAlchemy==2.0.27
typing_extensions==4.10.0
urllib3==2.2.2
virtualenv==20.26.1
waitress==3.0.0
Werkzeug==3.0.1
